requirements.txt
# Content anonymization dependencies
# Python >= 3.8

numpy>=1.20
pyyaml>=5.4
torch>=1.10
transformers>=4.30
sentencepiece>=0.1.99
scikit-learn>=1.0
sentence-transformers>=2.2
dtw-python>=1.1
tqdm>=4.60
openai>=1.0
einops>=0.6
